requirements.txt
# HTTP requests
httpx==0.27.0
requests==2.31.0

# YouTube Transcript (fallback)
youtube-transcript-api>=0.6.2

# Azure Storage
azure-storage-blob==12.19.0
azure-identity==1.15.0

# OpenAI and Azure OpenAI
openai==1.12.0

# Vector Store
chromadb==0.4.22
duckdb==0.9.2

# LangChain
langchain==0.1.0
langchain-openai==0.0.8
langchain-community==0.0.20

# UI
streamlit==1.32.0

# API
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
pandas==2.2.0
tenacity==8.2.3

